move skeleton to components
