+++ src/app/dashboard/stats/index.tsx
@@ -1,13 +1,11 @@
-import { ReactNode } from "react";
 import Link from "next/link";
-
-import { Skeleton } from "@/components/ui/skeleton";
+import { ReactNode } from "react";
 
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import AgeCard from "./age";
 import { AnimeCard, MangaCard } from "./animanga";
-import { DiscordCard } from "./discord";
 import { CodingCard } from "./coding";
+import { DiscordCard } from "./discord";
 import { SpotifyPlay } from "./spotify";
 
 export function ItemCard({
@@ -36,23 +34,5 @@ export function ItemCard({
     </Link>
   );
 }
-export function SkeletonCard({ count }: { count?: number }) {
-  const skeletons = Array.from({ length: count || 1 }).map((_, i) => (
-    <Card
-      className="hover:border-2 hover:border-black dark:hover:border-white"
-      key={i}
-    >
-      <CardHeader>
-        <CardTitle>
-          <Skeleton className="h-4 w-[100px]" />
-        </CardTitle>
-      </CardHeader>
-      <CardContent>
-        <Skeleton className="h-4 w-[250px]" />
-      </CardContent>
-    </Card>
-  ));
-  return <>{skeletons}</>;
-}
 
-export { AgeCard, AnimeCard, MangaCard, DiscordCard, CodingCard, SpotifyPlay };
+export { AgeCard, AnimeCard, CodingCard, DiscordCard, MangaCard, SpotifyPlay };
